Update script [ci skip]

+++ scripts/domain-check-2.sh
@@ -5,10 +5,27 @@
 #
 # Author: Matty < matty91 at gmail dot com >
 #
-# Current Version: 2.20
-# Last Updated: 12-Feb-2019
+# Current Version: 2.25
+# Last Updated: 24-Mar-2019
 #
 # Revision History:
+#
+#  Version 2.25
+#   Added support for .bid/.ng/.site domains -- https://github.com/hawkeye116477
+#
+#  Version 2.24
+#   Added support for .app/.top domains -- Vladislav V. Prodan <github.com/click0>
+#   Fixed support for .io/.xyz/.me/.pl/.ro domains -- Vladislav V. Prodan <github.com/click0>
+#
+#  Version 2.23
+#   Added support for .online domain -- https://github.com/hawkeye116477
+#
+#  Version 2.22
+#   Added support for .kz domains -- Vladislav V. Prodan <github.com/click0>
+#	Many thanks to the service that provided the API for the .KZ zone - https://www.ps.kz/
+#
+#  Version 2.21
+#   Fixed support for .pl domain -- https://github.com/hawkeye116477
 #
 #  Version 2.20
 #   Fixed support for .jp/.aero/.cn/.pl/.md/.tr/.it/.mx domains -- Vladislav V. Prodan <github.com/click0>
@@ -198,9 +215,11 @@ CUT=`which cut`
 GREP=`which grep`
 TR=`which tr`
 MAIL=`which mail`
+CURL=`which curl`
 
 # Place to stash temporary files
 WHOIS_TMP="/var/tmp/whois.$$"
+WHOIS_2_TMP="/var/tmp/whois_2.$$"
 
 #############################################################################
 # Purpose: Convert a date from MONTH-DAY-YEAR to Julian format
@@ -294,7 +313,7 @@ check_domain_status()
 {
     local REGISTRAR=""
     # Avoid WHOIS LIMIT EXCEEDED - slowdown our whois client by adding 3 sec
-    sleep 3
+    sleep 1
     # Save the domain since set will trip up the ordering
     DOMAIN=${1}
     TLDTYPE=$(echo ${DOMAIN} | ${AWK} -F. '{print tolower($NF);}')
@@ -331,6 +350,11 @@ check_domain_status()
     if [ "${TLDTYPE}" == "pl" ];
     then
        ${WHOIS} -h whois.dns.pl "${1}" | env LC_CTYPE=C LC_ALL=C ${TR} -d "\r" > ${WHOIS_TMP}
+    fi
+    if [ "${TLDTYPE}" == "kz" ];
+    then
+       ${CURL} -s "https://api.ps.kz/kzdomain/domain-whois?username=test&password=test&input_format=http&output_format=get&dname=${1}" \
+        | env LC_CTYPE=C LC_ALL=C ${TR} -d "\r" > ${WHOIS_2_TMP}
     fi
     # Parse out the expiration date and registrar -- uses the last registrar it finds
     REGISTRAR=`cat ${WHOIS_TMP} | ${AWK} -F: '/Registrar:/ && $2 != ""  { REGISTRAR=substr($2,2,17) } END { print REGISTRAR }' \
@@ -381,6 +405,9 @@ check_domain_status()
     elif [ "${TLDTYPE}" == "ua" ]; # added by @click0 20190212
     then
 	REGISTRAR=`cat ${WHOIS_TMP} | ${AWK} -F: '/registrar:/ && $2 != "" { REGISTRAR=substr($2,9,17) } END { print REGISTRAR }'`
+    elif [ "${TLDTYPE}" == "kz" ]; # added by @click0 20190223
+    then
+	REGISTRAR=`cat ${WHOIS_TMP} | ${AWK} -F": " '/Current Registar:/ && $0 != "" {print $2;}' | ${TR} -d " \r"`
     elif [ "${TLDTYPE}" == "cz" ]; # added by Minitram 20170830
     then
         REGISTRAR=`cat ${WHOIS_TMP} | ${AWK} -F: '/registrar:/ && $2 != "" { REGISTRAR=substr($2,5,17) } END { print REGISTRAR }'`
@@ -519,28 +546,6 @@ check_domain_status()
 		esac
             tday=`echo ${tdomdate} | ${CUT} -d'/' -f3`
 	    DOMAINDATE=`echo $tday-$tmonth-$tyear`
-    elif [ "${TLDTYPE}" == "me" ]; # for .me domain
-    then
-	tdomdate=`cat ${WHOIS_TMP} | ${AWK} '/Registry Expiry Date:/ { print $4 }'`
-	tyear=`echo ${tdomdate} | ${CUT} -d "-" -f 1`
-	tmon=`echo ${tdomdate} | ${CUT} -d "-" -f 2`
-               case ${tmon} in
-                     1|01) tmonth=jan ;;
-                     2|02) tmonth=feb ;;
-                     3|03) tmonth=mar ;;
-                     4|04) tmonth=apr ;;
-                     5|05) tmonth=may ;;
-                     6|06) tmonth=jun ;;
-                     7|07) tmonth=jul ;;
-                     8|08) tmonth=aug ;;
-                     9|09) tmonth=sep ;;
-                     10) tmonth=oct ;;
-                     11) tmonth=nov ;;
-                     12) tmonth=dec ;;
-                     *) tmonth=0 ;;
-               esac
-	tday=`echo ${tdomdate} | ${CUT} -d "-" -f 3 | ${CUT} -d "T" -f 1`
-	DOMAINDATE=`echo "${tday}-${tmonth}-${tyear}"`
     elif [ "${TLDTYPE}" == "ru" -o "${TLDTYPE}" == "su" ]; # for .ru and .su 2014/11/13
     then
            tdomdate=`cat ${WHOIS_TMP} | ${AWK} '/paid-till:/ { print $2 }'`
@@ -585,7 +590,29 @@ check_domain_status()
            esac
        tday=`echo ${tdomdate} | ${CUT} -d "-" -f 3 | ${CUT} -d "T" -f 1`
        DOMAINDATE=`echo "${tday}-${tmonth}-${tyear}"`
-    elif [ "${TLDTYPE}" == "com" -o "${TLDTYPE}" == "net" -o "${TLDTYPE}" == "org"  -o "${TLDTYPE}" == "link" -o "${TLDTYPE}" == "blog" -o "${TLDTYPE}" == "cafe" -o "${TLDTYPE}" == "biz" -o "${TLDTYPE}" == "us" -o "${TLDTYPE}" == "mobi" -o "${TLDTYPE}" == "tv" -o "${TLDTYPE}" == "co" -o "${TLDTYPE}" == "pro" -o "${TLDTYPE}" == "cafe" -o "${TLDTYPE}" == "in" -o "${TLDTYPE}" == "cat" -o "${TLDTYPE}" == "asia" -o "${TLDTYPE}" == "cc" -o "${TLDTYPE}" == "college" -o "${TLDTYPE}" == "aero" -o "${TLDTYPE}" == "online" ]; # added on 26-aug-2017 by nixCraft
+    elif [ "${TLDTYPE}" == "kz" ]; # for .kz @click0 2019/02/23
+    then
+           tdomdate=`cat ${WHOIS_2_TMP} | ${GREP} -A 1 "expire" | ${GREP} "utc" | ${AWK} -F\" '{print $4;}'`
+           tyear=`echo ${tdomdate} | ${CUT} -d'-' -f1`
+           tmon=`echo ${tdomdate} | ${CUT} -d'-' -f2`
+           case ${tmon} in
+                 1|01) tmonth=jan ;;
+                 2|02) tmonth=feb ;;
+                 3|03) tmonth=mar ;;
+                 4|04) tmonth=apr ;;
+                 5|05) tmonth=may ;;
+                 6|06) tmonth=jun ;;
+                 7|07) tmonth=jul ;;
+                 8|08) tmonth=aug ;;
+                 9|09) tmonth=sep ;;
+                 10) tmonth=oct ;;
+                 11) tmonth=nov ;;
+                 12) tmonth=dec ;;
+                 *) tmonth=0 ;;
+           esac
+       tday=`echo ${tdomdate} | ${CUT} -d "-" -f 3 | ${CUT} -d "T" -f 1`
+       DOMAINDATE=`echo "${tday}-${tmonth}-${tyear}"`
+    elif [ "${TLDTYPE}" == "com" -o "${TLDTYPE}" == "net" -o "${TLDTYPE}" == "org"  -o "${TLDTYPE}" == "link" -o "${TLDTYPE}" == "blog" -o "${TLDTYPE}" == "cafe" -o "${TLDTYPE}" == "biz" -o "${TLDTYPE}" == "us" -o "${TLDTYPE}" == "mobi" -o "${TLDTYPE}" == "tv" -o "${TLDTYPE}" == "co" -o "${TLDTYPE}" == "pro" -o "${TLDTYPE}" == "cafe" -o "${TLDTYPE}" == "in" -o "${TLDTYPE}" == "cat" -o "${TLDTYPE}" == "asia" -o "${TLDTYPE}" == "cc" -o "${TLDTYPE}" == "college" -o "${TLDTYPE}" == "aero" -o "${TLDTYPE}" == "online" -o "${TLDTYPE}" == "app" -o "${TLDTYPE}" == "io" -o "${TLDTYPE}" == "me" -o "${TLDTYPE}" == "xyz" -o "${TLDTYPE}" == "top" -o "${TLDTYPE}" == "bid" -o "${TLDTYPE}" == "ng" -o "${TLDTYPE}" == "site" ]; # added on 26-aug-2017 by nixCraft
     then
            tdomdate=`cat ${WHOIS_TMP} | ${AWK} '/Registry Expiry Date:/ { print $NF }'`
            tyear=`echo ${tdomdate} | ${CUT} -d'-' -f1`
@@ -655,7 +682,7 @@ check_domain_status()
 
     elif [ "${TLDTYPE}" == "pl" ] # NASK
     then
-          tdomdate=`cat ${WHOIS_TMP} | ${AWK} -F':' '/expiration date:/ || /renewal date:/ { print $2 }' | ${AWK} '{ print $1 ;}'`
+          tdomdate=`cat ${WHOIS_TMP} | ${AWK} -F: '/^expiration date:/ || /renewal date:/ { print $2; }' | ${AWK} -F" " '{ print $1; }'`
           tyear=`echo ${tdomdate} | ${CUT} -d'.' -f1`
           tmon=`echo ${tdomdate} | ${CUT} -d'.' -f2`
           case ${tmon} in
@@ -675,28 +702,6 @@ check_domain_status()
           esac
           tday=`echo ${tdomdate} | ${CUT} -d'.' -f3`
           DOMAINDATE=`echo "${tday}-${tmonth}-${tyear}"`
-    elif [ "${TLDTYPE}" == "xyz" ];
-    then
-        tdomdate=`cat ${WHOIS_TMP} | ${AWK} '/Registry Expiry Date:/ { print $4 }'`
-        tyear=`echo ${tdomdate} | ${CUT} -d "-" -f 1`
-        tmon=`echo ${tdomdate} | ${CUT} -d "-" -f 2`
-               case ${tmon} in
-                     1|01) tmonth=jan ;;
-                     2|02) tmonth=feb ;;
-                     3|03) tmonth=mar ;;
-                     4|04) tmonth=apr ;;
-                     5|05) tmonth=may ;;
-                     6|06) tmonth=jun ;;
-                     7|07) tmonth=jul ;;
-                     8|08) tmonth=aug ;;
-                     9|09) tmonth=sep ;;
-                     10) tmonth=oct ;;
-                     11) tmonth=nov ;;
-                     12) tmonth=dec ;;
-                     *) tmonth=0 ;;
-               esac
-        tday=`echo ${tdomdate} | ${CUT} -d "-" -f 3 | ${CUT} -d "T" -f 1`
-        DOMAINDATE=`echo "${tday}-${tmonth}-${tyear}"`
 
     elif [ "${TLDTYPE}" == "se" -o "${TLDTYPE}" == "nu" ];
     then
@@ -838,7 +843,7 @@ check_domain_status()
 
     elif [ "${TLDTYPE}" == "ro" ];	# added by nixCraft 07/jan/2019
     then
-        tdomdate=`cat ${WHOIS_TMP} | ${AWK} -F':' '/Expires On:/ { print $2 }'`
+        tdomdate=`cat ${WHOIS_TMP} | ${AWK} -F':' '/Expires On:/ { print $2; }' | ${AWK} '{ print $1; }'`
         tyear=`echo ${tdomdate} | ${CUT} -d "-" -f 1`
         tmon=`echo ${tdomdate} | ${CUT} -d "-" -f 2`
                case ${tmon} in
@@ -869,31 +874,7 @@ check_domain_status()
 
     elif [ "${TLDTYPE}" == "cn" ];	# for .cn @click0 2019/02/12
     then
-        tdomdate=`cat ${WHOIS_TMP} | ${AWK} -F':' '/Expiration Time:/ { print $2 }' | ${AWK} '{ print $1 ;}'`
-        tyear=`echo ${tdomdate} | ${CUT} -d "-" -f 1`
-        tmon=`echo ${tdomdate} | ${CUT} -d "-" -f 2`
-               case ${tmon} in
-                     1|01) tmonth=jan ;;
-                     2|02) tmonth=feb ;;
-                     3|03) tmonth=mar ;;
-                     4|04) tmonth=apr ;;
-                     5|05) tmonth=may ;;
-                     6|06) tmonth=jun ;;
-                     7|07) tmonth=jul ;;
-                     8|08) tmonth=aug ;;
-                     9|09) tmonth=sep ;;
-                     10) tmonth=oct ;;
-                     11) tmonth=nov ;;
-                     12) tmonth=dec ;;
-                     *) tmonth=0 ;;
-               esac
-        tday=`echo ${tdomdate} | ${CUT} -d "-" -f 3`
-        DOMAINDATE=`echo "${tday}-${tmonth}-${tyear}"`
-
-    elif [ "${TLDTYPE}" == "io" ];	# for .io @click0 2019/02/12
-    then
-        tdomdate=`cat ${WHOIS_TMP} | ${AWK} -F':' '/Registry Expiry Date:/ { print $2 }' \
-            | ${AWK} '{ print $1 ;}' | ${AWK} -FT '{print $1;}'`
+        tdomdate=`cat ${WHOIS_TMP} | ${AWK} -F':' '/Expiration Time:/ { print $2 }' | ${AWK} '{ print $1; }'`
         tyear=`echo ${tdomdate} | ${CUT} -d "-" -f 1`
         tmon=`echo ${tdomdate} | ${CUT} -d "-" -f 2`
                case ${tmon} in
@@ -1074,7 +1055,8 @@ fi
 echo
 
 ### Remove the temporary files
-rm -f ${WHOIS_TMP}
+[ -f "${WHOIS_TMP}" ] && rm -f ${WHOIS_TMP};
+[ -f "${WHOIS_2_TMP}" ] && rm -f ${WHOIS_2_TMP};
 
 ### Exit with a success indicator
 exit 0
